wip defining layers
